requirements.txt
aiosqlite==0.18.0
alembic==1.9.2
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
install==1.3.5
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.10.4
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
